requirements.txt
flask>=3.0.0
gunicorn>=23.0.0
databricks-sql-connector>=3.4.0
databricks-sdk>=0.33.0
pyyaml>=6.0
requests>=2.31.0
python-dotenv>=1.0.0
